cockpit: Add an example cockpit front end
This is meant as a POC to demonstrate how cockpit could be used as a
higher level web ui around mgmt.

We use a YAML file as the entry point, but that is just one API.

Please excuse my crappy javascript! JS hackers welcome!

TODO: Consider adding some sort of Status() API to mgmt for returning
information to ui's that might want to display it graphically.

XXX: Expect this feature branch to get rebased.