Check for tokens that have a missing colon
Colon added to a TODO that I've missed

Brought corrections to more "TODO" tokens that were missing a colon

Only do the check for a missing colon if the token is found inside a comment

One-liner comments, but I haven't seen multi-line comments in mgmt.

Added another colon that was forgotten

One more missing colon was forgotten. Fixed.