yaml2: Meta should keep defaults, and Res should have kind
This would previously panic since it wouldn't get a kind, and the meta
parameters would overwrite the defaults so it would block because limit
didn't have the default of +inf.

The removal of the SetKind was my fault in:

b8ff6938dfef73a252e45c5a74ca03436e121bad

It's funny because it ends in `bad`. Guess I should have checked that!