resources: aws: ec2: Close the listener after serve
The listener needs to be closed in order to prevent opening any new
connections after the server has been shut down. In those cases we
would panic because the http package would attempt to serve the client,
and send a message on a closed channel.

By closing the listener when we return from the server's goroutine, we
avoid this scenario, and can exit gracefully.