misc: Add github CONTRIBUTING support
Without this patch, there is no CONTRIBUTING.{md,txt,rst}
file.

This is a problem as it might not be obvious, for a first time
contributor, to check in .github/ for the PR template to know
how to contribute.

This fixes it by creating a CONTRIBUTING.txt file, which simply
points to the necessary existing documentation.