test: Fix commit testing
Currently, the testing of commits is in CI, and is CI
dependent.

This is a problem if we switch to a new CI, or if someone
wants to test locally.

This fixes it by making sure the test/test-commit-message
can be triggered locally, and finds the new commits.
As a side benefit, test/test-commit-message easier to understand:
It compares the HEAD of the branch until reaching a common
ancestor in origin, listing all the SHAs, and send those to
the testing functions.