lang: Fix ValueOf() with nil pointer values
Some forms of reflect.Value can cause ValueOf() to panic when there is a
nil pointer somewhere within the reflect.Value, whether that be a
container type like a struct, list or map, or just a raw nil pointer.

In these cases, ValueOf() attempted to dereference the pointer without
ever checking if it was nil. mgmt lang doesn't have pointers of any
kind, so the equivalent is the zero type of the value. In such cases,
gracefully attempt to handle by constructing a types.Value based on a
new instance of the type of the pointer reflect.Value instead.

Signed-off-by: Joe Groocock <me@frebib.net>