lib: Move to go modules with regret
I'm incredibly annoyed that go modules are being forced onto us, as the
old system with vendor/ and git submodules worked great. Unfortunately,
so much FUD around git submodules, even by well-known, highly paid
engineers has caused this solution to be deemed most unacceptable.

So much for the golang compatibility promise-- turns out it doesn't
apply to the useful parts that we actually care about like this.

Thanks to frebib for his help with this suffering.