test: Add a golanglint-ci test
This adds the basic plumbing and tests so that this runs both in our
normal shell based test suite, but also with github actions when it
happens to be running there. The advantage of that is that users will
get a visual output of the issues that we find.

This also fixes a small durationcheck bug.