lang: remove Expr::SetValue()
The method is used by the function engine to store the latest value
which was emitted by Func::Stream(), but this is redundant because the
engine also stores that value in the engine's own table[vertex].

Moreover, most of the SetValue() implementations have a comment
indicating that the function is a no-op, or is never called and is only
implemented for consistency.

Expr::Value() is still useful because it is used during unification. For
example, printf("%d", x) can tell the type checker that x must be an
int, but only if the format argument's value is statically known. So it
is still useful to have a method for asking for that value if it is
known.