lang: ast, interfaces, interpret: Pass through table
This is an alternate method for getting values from the AST. Instead of
running SetValue on each relevant AST node, we instead store the
function pointers that will need values, and then pass in a table
mapping those pointers to the actual value.

The downside is that the code is a little less obvious and it's longer.
We do a table lookup instead of simply calling Value(). The upside is
that it's easier to test, and also natively thread-safe, since we don't
need to call SetValue anywhere. Lastly, because we don't call SetValue,
there is a small chance an inconsistency bug could fall-through the
SetValue checks, but that's unlikely and not a design requirement.