wrap func declarations in a ConstFunc
Function declarations produce a graph containing a node which transforms
input values to output values, but CallFunc expects a FuncValue instead.
We thus wrap function declarations in a node with in-degree zero, which
produces a FuncValue, which generates a sub-graph containing the node
which transforms input values to output values.