using new Txn.Erase() instead of double-commit
turns out the semantics of Txn.Reverse() are _not_ to reverse all the
changes made by the last transaction, but to reverse all the changes
made by the last _non-empty_ transaction. Txn.Erase() makes it explicit
that we do not want the just-committed transaction to be reversed by the
next call to Txn.Reverse().