XXXDAGEXXX lang: funcs: Ensure we only send complete structs
We were silently sending a struct with a missing field. This could
happen if we had the same field in two places, such as $x + $x. As a
result this was giving the wrong answer of `2` for an equation like
`2 + 2`, because one of the values was the default value of zero.