lang: Add a new DAG function engine (dage)
This adds a new implementation of the function engine that runs of the
DAG function graph. This version is notable in that it can run a graph
that changes shape over time. To make changes to the shape of the graph,
you must use the new transaction (Txn) system. This system implements a
simple garbage collector (GC) for scheduled removal of nodes that the
transaction system "reverses" out of the graph.

Lastly, this adds some associated plumbing and other fixtures to make
all of this work.