lang: ast: split ExprPoly in two
Previously, all top-level expressions were polymorphic, so ExprPoly was
both capturing the top-level scope and also acting as a sentinel so that
ExprVar and ExprCall know that they need to make their own copy of
ExprPoly.Definition.

Now, all only some top-level definitions are polymorphic, but they all
need to capture the top-level scope, so ExprPoly is now solely focused
on being a sentinel while the new ExprTopLevel is solely focused on
capturing the top-level scope.