engine: graph, util: Restore send/recv behaviour
A regression in 4b0cdf912372a6da76fea3130b6a5b3b59c428e7 caused the
basic send/recv functionality to break for simple scenarios. This was
due to inadequate testing, and a partial misunderstanding of the
situation.

New testing should hopefully catch more cases, but send/recv and
compile-time checks are still not as complete as is probably possible.