engine: resources: Rewrite MountRes to rely on systemd units over fstab
The mount approach was flawed. It relied on editing fstab, then restarting the
systemd service that is responsible for mounting all defined filesystems. Among
other issues, this approach made it impossible for mgmt to unmount managed
filesystems, because after removing them from fstab, they would just be ignored
by systemd.

The new approach is based on creating a systemd unit for each managed mount,
through a unit file in /etc/systemd/system. These files are managed by wrapped
File resources now. Then a wrapped Svc resource manages the lifecycle of the
mount. This works even after removing the unit definition from systemd, because
as long as the filesystem is still mounted, systemd will be able to derive a
transient unit that can be managed.

This also simplifies the logic of the MountRes a lot, because we don't have to
be so careful anymore about matching existing fstab entries to what is expected
by a given Mount resource. We just overwrite the unit definition with whatever
is specified, and rely on systemd to handle the details. Systemd will even
create mount points if they don't yet exist.

Currently this approach is not yet equipped to handle the case that a mount
could be defined via fstab already, or through a unit file that is in a
different location than /etc/systemd/system. As such, this second iteration
should be considered as a partial (if more rounded than the first one)
solution.

Fixes #525